backend/requirements.txt

fastapi
uvicorn
pillow
torch
transformers
numpy
python-multipart
torchvision